LFO support in waves :

+++ wave.h
@@ -12,20 +12,20 @@ typedef struct wave_s{
 
   /*wavetable properties*/
   
-  size_t F;
+  float F;
   size_t A;
-  size_t STEP; //= w->T / w->F * SRATE;
-  size_t WAIT; //= 1.0 / STEP (> 1 if STEP < 1
+  float STEP; //= w->T / w->F * SRATE;
+  short WAIT; //= 1.0 / STEP (> 1 if STEP < 1
   size_t cur;
 
 } wave;
 
 sample wave_next(void *);  
 void wave_init(wave **w);/*create a wave structure*/
-void wave_freq(wave *w, size_t f);
+void wave_freq(wave *w, float f);
 void wave_setwavetable(wave **w, wavetable *wt);
 void wave_mkwtab(wave *w, wavetable *wt);
 void mod_mkwave(module *m, wave *w);/*set m->mod = new wave; set mod->next = wave_next*/
-module *wave_new( wavetable *wtab, unsigned freq );
+module *wave_new( wavetable *wtab, float freq );
 
 #endif
